Task2

+++ Tyuiu.ErmakovAA.Sprint3.Task2.V18.Lib/DataService.cs
@@ -8,10 +8,13 @@ namespace Tyuiu.ErmakovAA.Sprint3.Task2.V18.Lib
             double multiplySeries = 1;
             do
             {
-                multiplySeries = multiplySeries * ((Math.Cos(value) + Math.Pow((startValue / 8), 3)));
+                multiplySeries = multiplySeries * (Math.Cos(value) + (Math.Pow(((double)startValue / 8), 3)));
                 startValue++;
             } while (startValue <= stopValue);
             return Math.Round(multiplySeries, 3);
         }
     }
 }
+
+
+
